fix(security): address PR review feedback
