+++ apps/hyperlocalise-web/src/lib/agents/github/bot.ts
@@ -2,7 +2,6 @@ import type { GitHubAdapter, GitHubRawMessage } from "@chat-adapter/github";
 import { createGitHubAdapter } from "@chat-adapter/github";
 import { Chat, emoji } from "chat";
 import type { Message, Thread } from "chat";
-import { getInstallationOctokit } from "@/lib/agents/github/app";
 
 import {
   createHyperlocaliseAgent,
@@ -22,6 +21,7 @@ import type { GitHubFixRequestedEventData, GitHubFixQueue } from "@/lib/workflow
 
 import { parseFixCommand } from "./commands";
 import { buildFixEvent } from "./events";
+import { requesterCanRunFix } from "./permissions";
 import { createGitHubFixTools } from "./tools";
 
 type GitHubBotOptions = {
@@ -43,53 +43,6 @@ async function getOrganizationIdByInstallationId(installationId: string) {
     .where(eq(schema.githubInstallations.githubInstallationId, installationId))
     .limit(1);
   return installation?.organizationId ?? null;
-}
-
-async function requesterCanRunFix(event: GitHubFixRequestedEventData) {
-  let data: { permission?: string };
-  try {
-    const octokit = await getInstallationOctokit(event.installationId);
-    ({ data } = await octokit.rest.repos.getCollaboratorPermissionLevel({
-      owner: event.repositoryOwner,
-      repo: event.repositoryName,
-      username: event.trigger.requesterLogin,
-    }));
-  } catch (error) {
-    if (isPermissionLookupDenial(error)) {
-      return false;
-    }
-    throw error;
-  }
-
-  return (
-    data.permission === "admin" || data.permission === "maintain" || data.permission === "write"
-  );
-}
-
-function isPermissionLookupDenial(error: unknown) {
-  if (typeof error !== "object" || error === null || !("status" in error)) {
-    return false;
-  }
-  const status = error.status;
-  if (status === 404) {
-    return true;
-  }
-  if (status !== 403) {
-    return false;
-  }
-  const message = "message" in error && typeof error.message === "string" ? error.message : "";
-  const response =
-    "response" in error && typeof error.response === "object" && error.response !== null
-      ? error.response
-      : null;
-  const headers =
-    response && "headers" in response && typeof response.headers === "object"
-      ? response.headers
-      : null;
-  const rateLimitRemaining =
-    headers && "x-ratelimit-remaining" in headers ? headers["x-ratelimit-remaining"] : null;
-
-  return rateLimitRemaining !== "0" && !/rate limit/i.test(message);
 }
 
 function buildGitHubFixInstructions(event: GitHubFixRequestedEventData) {
